Migrate to-do list from markdown to plain text
I got tired of using a todo.md.  Markdown is great, but not expressive
enough for a to-do list.  Thanks to @todotxt for changing me (lmao).

This had the immediate consequence of me renaming ~/Documents/Markdown
to ~/Documents/Notes, which is much better tbh.  I also reset
`textwidth` in Vim for all todo.txt files, since it strictly requires
one task per line.